refactor: move help functions to phase_utils.py
